Fix checkmates
